requirements.txt
python-telegram-bot==20.7
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
schedule==1.2.0
python-dotenv==1.0.0
aiohttp==3.8.5
python-dateutil==2.8.2
pillow==10.0.0
seaborn==0.12.2
ta
setuptools>=68.0
wheel
build
